backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
python-engineio==4.8.0
python-socketio==5.10.0
eventlet==0.33.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0